% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRfactor.R
\name{addRfactor}
\alias{addRfactor}
\title{addRfactor.R}
\usage{
addRfactor(cathTable, RfactorDF)
}
\arguments{
\item{cathTable}{(data.frame) A dataframe containing domain entries 
from the CATH database}

\item{RfactorDF}{(data.frame) A dataframe containing the R factors and
the matching protein domains}
}
\value{
The \code{cathTable} with the Rfactors added to each protein 
entry
}
\description{
\code{addRfactor} adds the Rfactors including
\code{Rfree}, \code{RWork}, and \code{RObserved} to the corresponding
domains in the CATH table
}
