% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addValidation.R
\name{addValidation}
\alias{addValidation}
\title{addValidation.R}
\usage{
addValidation(xmlFile, pdbID, cathTable)
}
\arguments{
\item{xmlFile}{(vector) A XML file path that will be read}

\item{pdbID}{(string) The ID of the PDB ID}

\item{cathTable}{(data.frame) A dataframe containing domain entries 
from the CATH database}
}
\value{
The \code{cathTable} with the validation scores added to 
each protein entry
}
\description{
\code{addValidation} adds the stereochemical validation reports
into the \code{cathTable} by reading the given XML file.
This includes the \code{clashscore}, \code{Ramachandran Score},
\code{RSRZ score} and \code{percent outliers}.
}
