% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCathData.R
\name{fetchCathData}
\alias{fetchCathData}
\title{fetchCathData.R}
\usage{
fetchCathData()
}
\value{
The path to the downloaded CATH chain list text file containing
the CATH hierarchy data
}
\description{
\code{fetchCathData} fetches the CATH chain list from the 
CATH database in the form of a textfile
}
