% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCathlist.R
\name{fetchCathList}
\alias{fetchCathList}
\title{fetchCathlist.R}
\usage{
fetchCathList()
}
\value{
The path to the downloaded CATH chain list text file containing
the CATH hierarchy data
}
\description{
Fetch the CATH chain list from the CATH database in the form
of a textfile
}
