% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchRFactor.R
\name{fetchRFactor}
\alias{fetchRFactor}
\title{fetchRFactor.R}
\usage{
fetchRFactor(cathTable)
}
\value{
A dataframe RFactorDF that contains the R facotr vlaues for each PDB domain
}
\description{
Fetch the R Facotr values from RestfulWebservices from the PDB
}
