% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixMissingRValues.R
\name{fixMissingRValues}
\alias{fixMissingRValues}
\title{fixMissingRValues.R}
\usage{
fixMissingRValues(cathTable)
}
\arguments{
\item{cathTable}{(data.frame) Data frame containing annotated cath data
that will have Rfactor values replaced}
}
\value{
(data.frame) The \code{cathTable} with "null" R factor
values replaced
}
\description{
\code{fixMissingRValues} Replace "null" R factor values with
estimated R factor values in the CathTable
}
\details{
\code{fixMissingRValues} Outputs the \code{cathTable} with
missing Rfactor values replaced if at least one of the
\code{RObserved}, \code{RWork}, or \code{RFree} is not "null".
If no R factors are available, they will be left as \code{NA}.
}
