% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCathKeyDF.R
\name{generateCathKeyDF}
\alias{generateCathKeyDF}
\title{generateCathKeyDF.R}
\usage{
generateCathKeyDF(cathTable)
}
\arguments{
\item{cathTable}{(data.frame) A dataframe that contains the CATH data. Can
be generated by calling the \code{generateCathTable()} function.}
}
\value{
A dataframe containing the hierarchial data of the CATH database
and the corresponding descriptions to each domain.
}
\description{
\code{generateCathKeyDF} is a function in the Nh3D2.0 package. It generates
 a hierarchial database containing all CATH domains and descriptions for
 the user as a reference.
}
\examples{
\dontrun{

# Generate the cathTable
cathTable <- generateCathTable()

# Make the key dataframe
cathKeyDF <- generateCathKeyDF(cathTable)

(cathKeyDF)

}

}
