% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCathTable.R
\name{generateCathTable}
\alias{generateCathTable}
\title{generateCathTable.R}
\usage{
generateCathTable()
}
\value{
A dataframe containing the CATH domains and some annotations including
the group identifiers (class, architecture, topology and homology), protein
sequence length, and resolution.
}
\description{
\code{generateCathTable} is a function in the Nh3D2.0 package. It generates
a data frame that contains the entire CATH domains and their entries. 
This \code{cathTable} is required in order to run the \code{Nh3D} function.
}
\examples{
\dontrun{

# Generate the cathTable
cathTable <- generateCathTable()

(cathTable)

}

}
