% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestDomains.R
\name{getBestDomains}
\alias{getBestDomains}
\title{getBestDomains.R}
\usage{
getBestDomains(cathTable, n = 10)
}
\arguments{
\item{cathTable}{(data.frame) A data frame that contains 
the processed cath data from the \code{Nh3D} function.}

\item{n}{(numeric) The number of representatives the 
user would like to retrieve.}
}
\value{
A dataframe containing the select best representatives
within a CATH identifier based on their ranking.
}
\description{
\code{getBestDomains} retrieves the n best representative 
domains from a provided cathTable that have been ran with
\code{Nh3D}.
}
\examples{
\dontrun{

# We first generate a cathTable in our local global env.
cathTable <- generateCathTable()
 
# We then select a topology group that we would like to retrieve a dataset of
# non-homologous proteins for. 

class <- 1
architecture <- 10
topology <- 220

# Run Nh3D with the parameters
myTable <- Nh3D(cathTable, class  = class, architecture = architecture, 
topology = topology)

# Get the best domains
getBestDomains(myTable, 4)

}

}
