% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCathList.R
\name{readCathList}
\alias{readCathList}
\title{readCathList.R}
\usage{
readCathList(filepath)
}
\arguments{
\item{filepath}{A character vector containing the filepath to the 
CATH data as a form of text}
}
\value{
The CATH data in the form of a dataframe
}
\description{
Read the CATH list from the provided filepath and convert
it into a datatable
}
