% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCathNames.R
\name{readCathNames}
\alias{readCathNames}
\title{readCathNames.R}
\usage{
readCathNames()
}
\value{
The names, CATH identifier and description of each CATH 
protein domain.
}
\description{
\code{readCathNames} reads the \code{cath-names.txt} file from the 
 CATH database to create a dataframe containing the names of each 
 protein domain in CATH and the corresponding descriptions and 
 CATH identifier.
}
