% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineCathTable.R
\name{refineCathTable}
\alias{refineCathTable}
\title{refineCathTable.R}
\usage{
refineCathTable(cathTable)
}
\arguments{
\item{cathTable}{A dataframe containing the modified cath hierarchy data}
}
\value{
The refined and modified cathTable only containing domains that have
a length greater than 50bp and a resolution less than 2.2A.
}
\description{
Remove any domains in the cathTable that have a domain length less than 50
and a resolution value greater than 2.2A.
}
