% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLowestResolution.R
\name{removeLowestResolution}
\alias{removeLowestResolution}
\title{removeLowestResolution.R}
\usage{
removeLowestResolution(cathTable)
}
\arguments{
\item{cathTable}{(data.frame) A dataframe containing domain entries 
from the CATH database}
}
\value{
The \code{cathTable} with the lowest resolution scores 
in each homology group removed.
}
\description{
\code{removeLowestResolution} removes proteins entries that have
a resolution score lower than the mean in the homology group it
belongs to.
}
