% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCath.R
\name{updateCath}
\alias{updateCath}
\title{updateCath.R}
\usage{
updateCath()
}
\value{
The \code{cathTable} with the updated entries
}
\description{
\code{updateCath} updates and/or creates the CATH data and 
the data frame containing the cath data.
}
\examples{
\dontrun{

# Generate a CathTable
cathTable <- generateCathTable()

# We want to update the table, so 
# we call the function
cathTable <- updateCath()

# You can also call this function
# to create the cathTable

}

}
