% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight}
\alias{weight}
\title{weight.R}
\usage{
weight(cathTable)
}
\arguments{
\item{cathTable}{(data.frame) A dataframe containing domain entries 
from the CATH database}
}
\value{
The \code{cathTable} with the ranks of each protein entry in each
homology group added.
}
\description{
\code{weight} weighs the Resolution, Rfactor and
Stereochemical validation scores for each protein entry in 
each homology of the provided \code{cathTable}. Weighting 
is based off the multi-objective optimization method in 
finding the optimal values in statistics.
}
