% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCathCat.R
\name{addCathCat}
\alias{addCathCat}
\title{addCathCat.R}
\usage{
addCathCat(cathTable)
}
\arguments{
\item{cathTable}{(data.frame) A dataframe containing the
modified cath hierarchy data}
}
\value{
The modified dataframe containing the keys to the hierarchy
}
\description{
\code{addCathCat} adds to the cathTable dataframe the CATH 
hierarchy key of each protein homology in the cathtable
}
