% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addValidation.R
\name{addValidation}
\alias{addValidation}
\title{addValidation.R}
\usage{
addValidation(xmlFile, pdbID, cathTable)
}
\arguments{
\item{xmlFile}{A XML file path}

\item{cathTable}{The dataframe containing CATH data}

\item{ID}{The ID of the PDB ID}
}
\description{
Read and add the validation report numbers
into the cathTable and return it.
}
