% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchRFactor.R
\name{fetchRFactor}
\alias{fetchRFactor}
\title{fetchRFactor.R}
\usage{
fetchRFactor(cathTable)
}
\arguments{
\item{cathTable}{(data.frame) A dataframe containing domain entries 
from the CATH database}
}
\value{
The \code{cathTable} with the Rfactors added to each protein 
entry
}
\description{
\code{fetchRfactor} retrieves the R factor values for each protein entry in
the provided \code{cathTable}. The function calls \code{addRfactor} in order
to extract and add XML components into the \code{cathTable}.
}
