% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchVerifyData.R
\name{fetchVerifyData}
\alias{fetchVerifyData}
\title{fetchVerifyData.R}
\usage{
fetchVerifyData(cathTable)
}
\arguments{
\item{cathTable}{The dataframe containing the CATH data
that will me modified}
}
\description{
Fetch verification data from the 
wwPDB site by downloading the XML file containing
stereochemical validation report data
}
