% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nh3D.R
\name{Nh3D}
\alias{Nh3D}
\title{Nh3D.R}
\usage{
Nh3D()
}
\description{
The main function to Nh3D2.0
Calling this function will generate a dataset with the best representatives
for each topology group where all homologies in each group are non-homologous.
}
