% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nh3D.R
\name{Nh3D}
\alias{Nh3D}
\title{Nh3D.R
\code{Nh3D} is the main function to Nh3D2.0.
Calling this function will generate a dataset with the best representatives
for each topology group where all homologies in each group are non-homologous.
It calls several helper functions in order to generate the database.}
\usage{
Nh3D(cathTable, class, architecture = NULL, topology = NULL,
  homology = NULL)
}
\arguments{
\item{cathTable}{(data.frame) A dataframe that contains the CATH data. Can
be generated by calling the \code{generateCathTable()} function.}

\item{class}{(numeric) The class identifier of a class of interest. The
class identifier must be provided to reduce processing times.}

\item{architecture}{(numeric) The architecture identifier in the CATH
database. Automatically set to NULL.}

\item{topology}{(numeric) A topology identifier from the CATH database. 
Automatically set to NULL.}

\item{homology}{(numeric) A homology identifier from the CATH database. 
Automatically set to NULL.}
}
\value{
A dataframe containing the best representative PDB domains in a specified
CATH group according to the CATH database.
}
\description{
Nh3D.R
\code{Nh3D} is the main function to Nh3D2.0.
Calling this function will generate a dataset with the best representatives
for each topology group where all homologies in each group are non-homologous.
It calls several helper functions in order to generate the database.
}
\examples{
\dontrun{
 
# We first generate a cathTable in our local global env.
cathTable <- generateCathTable()
 
# We then select a topology group that we would like to retrieve a dataset of
# non-homologous proteins for. 

class <- 1
architecture <- 10
topology <- 220

# Run Nh3D with the parameters
myTable <- Nh3D(cathTable, class  = class, architecture = architecture, 
topology = topology)

(myTable)

}

}
