% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCathData.R
\name{readCathData}
\alias{readCathData}
\title{readCathData.R}
\usage{
readCathData(filepaths)
}
\arguments{
\item{filepaths}{(vector) A vector containing the filepaths to the CATH data files.}
}
\value{
The CATH data in the form of a dataframe
}
\description{
\code{readCathData} reads the CATH list from the provided filepath and converts
it into a dataframe
}
